+++ examples/data-sources/morpheus_cloud_datastore/data-source.tf
@@ -1,0 +1,4 @@
+data "morpheus_cloud_datastore" "morpheus_datastore" {
+  name     = "datastore01"
+  cloud_id = 4
+}