+++ examples/data-sources/morpheus_environments/data-source.tf
@@ -1,0 +1,3 @@
+data "morpheus_environments" "tf_environments" {
+  sort_ascending = false
+}