+++ morpheus/resource_azure_cloud.go
@@ -192,6 +192,9 @@ func resourceAzureCloud() *schema.Resource {
 				Computed:     true,
 			},
 		},
+		Importer: &schema.ResourceImporter{
+			StateContext: schema.ImportStatePassthroughContext,
+		},
 	}
 }