+++ CHANGELOG.md
@@ -7,6 +7,11 @@ NOTES:
 * Updated the `terraform-plugin-docs` library from v0.16.0 to v0.19.4
 * Updated the `terraform-plugin-sdk` library from v2.31.0 to v2.34.0
 * Fixed an issue in which the `role_ids` attribute for the `morpheus_user` resource was indicating a change when a change should not have been detected. [250](https://github.com/gomorpheus/terraform-provider-morpheus/issues/250)
+* Added support for the `morpheus_cloud_datastore` data source to lookup a cloud's datastore. [231](https://github.com/gomorpheus/terraform-provider-morpheus/issues/231)
+
+FEATURES:
+
+* **New Data Source:** `morpheus_cloud_datastore`
 
 ## 0.9.10 (July 16, 2024)