+++ CHANGELOG.md
@@ -11,12 +11,16 @@ NOTES:
 * Added support for the `morpheus_storage_volume` data source to lookup storage volumes. [230](https://github.com/gomorpheus/terraform-provider-morpheus/issues/230)
 * Added support for the `morpheus_environments` data source to lookup environments and return a list of environment ids. [233](https://github.com/gomorpheus/terraform-provider-morpheus/issues/233)
 * Added support for the `morpheus_groups` data source to lookup groups and return a list of group ids. [233](https://github.com/gomorpheus/terraform-provider-morpheus/issues/233)
+* Added support for the `morpheus_tasks` data source to lookup tasks and return a list of task ids. [233](https://github.com/gomorpheus/terraform-provider-morpheus/issues/233)
+
 
 FEATURES:
 
 * **New Data Source:** `morpheus_cloud_datastore`
 * **New Data Source:** `morpheus_storage_volume`
 * **New Data Source:** `morpheus_environments`
+* **New Data Source:** `morpheus_groups`
+* **New Data Source:** `morpheus_tasks`
 
 ## 0.9.10 (July 16, 2024)