+++ CHANGELOG.md
@@ -14,6 +14,7 @@ NOTES:
 * Added support for the `morpheus_tasks` data source to lookup tasks and return a list of task ids. [233](https://github.com/gomorpheus/terraform-provider-morpheus/issues/233)
 * Added support for the `morpheus_policies` data source to lookup policies and return a list of policy ids. [233](https://github.com/gomorpheus/terraform-provider-morpheus/issues/233)
 * Added support for the `morpheus_user_groups` data source to lookup user groups and return a list of user group ids. [233](https://github.com/gomorpheus/terraform-provider-morpheus/issues/233)
+* Added support for the `morpheus_tenants` data source to lookup tenants and return a list of tenant ids. [233](https://github.com/gomorpheus/terraform-provider-morpheus/issues/233)
 
 FEATURES:
 
@@ -24,6 +25,7 @@ FEATURES:
 * **New Data Source:** `morpheus_tasks`
 * **New Data Source:** `morpheus_policies`
 * **New Data Source:** `morpheus_user_groups`
+* **New Data Source:** `morpheus_tenants`
 
 ## 0.9.10 (July 16, 2024)